+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Bugs fixed
 * Allow input fasta files with alternative standard extensions.
 * Fixed handling of reference annotations that may not have /gene qualifiers for all annotations.
+* Fixed handling of input genome when it's the same as the reference.
 
 ## [Version 1.3.1](https://gitlab.com/LPCDRP/hybran/-/tags/1.3.1)